Organised Code
Creating MVC structure in folder
